refactor: 기존 MenuProduct, MenuProductDao 제거
