+++ src/main/java/kitchenpos/domain/Menu.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import kitchenpos.menu.domain.MenuProduct;
+
 import java.math.BigDecimal;
 import java.util.List;
