+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -3,5 +3,5 @@ package kitchenpos.menu.domain;
 import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<MenuGroup, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 }
